dts: add initial device tree for tiny-ex1 platform.
Add a Device Tree for the Tiny-EX1 platform configuration. This is an
experimental platform meant for testing various configurations and
system layouts.

Eventually, we'll come up with some fixed platform configurations. The
Tiny-EX1 is as follows:

 * CPU: 4x ARM Neoverse N1 cores, 2x Clusters (2x2)
 * RAM: 512M
 * UART: 2x non-secure PL011
 * Misc: ARM GICv3, Generic Timer

PLEASE NOTE: This is simply a name for a particular QEMU system
configuration, just to emulate a real-life system. There are NO
modifications made to QEMU, this is just a device tree for a made-up
platform.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>