pmap: introduce kernel pmap interface
The pmap interface in the kernel is responsible for handling a number of
things:
 * interaction between the kernel and the mmu/mmu system registers.
 * handling physical memory maps, called pmap_t
 * handling pagetables

The pmap interface is the lowest-level memory maangement interface. The
virtual memory (vm) interface will be built on-top of pmap. This commit
adds the initial pmap.h, along with refactoring the logic previously in
arm_vm_init into the new pmap.c.

The pagetables_region is now managed by pmap, using the pmap_ptregion
API, and the root kernel pagetable is now pointed to by kernel_tte,
which is also defined in pmap.h

arm_vm_init is now only responsible to verifying the memory information
passed to it via tBoot, and then invoking the pmap_ptregion api to
allocate some pages within the pagetables_region, and then allocate
virtual memory for the GIC, UART and Kernel.

Future commits will extend pmap, and adding the vm_map_t structure. The
only virtual memory maps that are created manually are the three in
arm_vm_init - everything else MUST be created via vm_map_create().

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>