pmap: remove kernel_tte allocation from pmap_ptregion_create
Calling pmap_ptregion_create no longer allocates a physical page in the
pagetables region for the kernel. Instead, this is now left to the caller.
This is because of the future introduction of an 'invalid_tte', which is
set to a translation table base register which isn't being used, i.e.
TTBR0_EL1 before userspace is configured.

Additionally, the physical address of the pagetables region is now stored
as a private global in pmap.c, as the virtual addresses may change.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>