arch: assign exception vector after bootstrap tables are configured.
Currently we're setting the _ExceptionVectorBase to VBAR_EL1 before the
bootstrap pagetables are setup. This means that once we invalidate
TTBR0_EL1 (once we're off the bootstrap tables), there is no longer a
1:1 mapping for physical-to-virtual addresses.

Essentially, when we enter the exception vector there is an immediate
data abort because VBAR_EL1 is set to a physical address - and we're
completely off physical addressing now.

So, to solve this we just move the assignment of VBAR_EL1 until after
the MMU is enabled, and we convert the physical address of
_ExceptionVectorBase to a KVA, like how we do with kernel_init.

Signed-off-by: Harry Moulton <me@h3adsh0tzz.com>