kern: refactor the entire kernel
This is essentially a complete re-write of the kernel, using a different build-system (make) and fixing a number of strange issues with the previous version. The latest kernel version as-of this commit is tinykern-0190.32.441.2.

Some things that are still missing:
* much documentation, for example the boot process
* sources for tinyBoot, these still need to remain closed-source for multiple reasons
* automated build version incrementing.
* boot.py script - i'll add this back soon.
* probably some other bits i've forgotten.

An updated TODO with core kernel functionality i want to work on in the short-term is included.