vm: temporary fix for vm_map_entry virtual address issue
An issue exists with vm_map and vm_map_entry where some entries,
particularly the second one, gives virtual addresses that are
significantly off - meaning that they aren't aligned with the previous
entry.

The way vm_map allocations work at the moment is that all virtual
address allocations must be contiguous. In the future there will be a
way to map virtual memory anywhere within the bounds described in the
vm_map structure, but for now they must be contiguous.

This adds a temporary workaround for when we're allocating memory for
the kernel task structure. An allocation of 0 bytes is made, which is
rounded up to VM_PAGE_SIZE.