task: add initial kernel task interface
Add the initial kernel task interface and task_t structure. Currently
this is quite limited, and the schedular hasn't been implemented yet.

The initial idea here is that the task.h has a pointer to the kernel
task, the current task, and a general list of all tasks on the system. A
task_t contains a process id (I'm not thinking about a process layer
on-top of tasks yet), whether the task is actively executing or not, the
cpu_context, vm_map, timings (total time the task has been executing,
and the time it's been executing since the schedular last switched), adn
it's priority.

The tasks are stored as a doubly-linked list, so each task stores a
pointer to the one before and after itself.

NOTE:

As I'm writing this I'm realising that makes the task list rather
redundant, and instead we could just have the pointer to the kernel task
(which is essentially the root) and the current task. This change will
be made in a future commit.