arch: basic context switcher - __fork64_switch
This change adds a basic context switcher, called __fork64_switch. This
takes a pointer to two task_t's: the current, and the target.

The current callee register state (x19-x31) is saved to the current
task_t (x0), and then the context from the target (x1) is restored to
the x19-x31 registers.

When __fork64_switch returns, it should be back in the target task. Some
things to consider here, however:

 * additional work needs to be done so task switches work across virtual
   memory spaces (kernel/user).
 * other structures need to be updated during the switch, should this
   happen before or after the call to __fork64_switch??
 * probably some other bits i've not thought of yet.