test: experimental task switching test
This is a very early attempt at switching between two different task_t
structures with different cpu contexts. The process here is essentially:

 * initialise the task interface with task_init (does nothing atm)
 * create the kernel_task, passing an entry point and the kernel vm_map.
 * invoke __fork64_switch, passing the kernel task as the current and
   target, essentially switching from itself to itself. This is to check
   that the task structure is not corrupted.
 * print out the current state of registers x19-x31.
 * create a test_task structure, and copy the kernel_task context to it,
   but modify the program counter to point to the kernel task entry
 * invoke __fork64_switch again, with kernel_task as current, and
   test_task as target

After the __fork64_switch has completed, we should be whereever the
test_target pc is pointing to. This should output a simple message to
make it clear that it's switch to a different point. It will then invoke
__fork64_switch again to swap back to the original program flow, where
it prints that the kernel init has completed.

Some things to think of here:

 * The context is the same between the two tasks, apart from the program
   counter. What if we change it slightly? what if the lower registers
   are different? does that cause any issues?