machine/timer: add machine wrapper for architecture timer
Add a machine-layer wrapper for the architecture timer. This isn't fully
abstracted yet, the timer code could arguably be moved in drivers/
rather than arch/, but since it's a arm-specific thing, and the kernel
only supports arm, it should be fine for now.

Some reset values are also defined in the header, as well as the
interrupt IDs for the EL1 physical and virtual interrupts. If we end up
supporting other timer drivers, such as the one on the raspberry pi,
then this can be made more generic.