kern: handle EL1 physical timer interrupt
When the kernel receives a physical timer interrupt at EL1 (IRQ30) it
needs to reset the timer. For now, this will be done with the default
reset value, so every time the timer fires, it resets with the same
interval.

This timer will be used for invoking the schedular. I'd like to look at
how the schedular could set the time each cycle, maybe with a global
somewhere?