exception: fixup misaligned registers in exception dump
When an exception occurs, the register state (in particular the
exception registers, i.e. ELR, ESR, FAR) did not contain the correct
values. The exception frame structure has been adjusted, and the offsets
to which registers are saved updated.

This contains some code from an upcoming patch for tasks, so this won't
build in it's current state.