tasks: improve initial task interface
This makes some big changes to the task interface and the design for it.
Tasks no longer directly point to eachother, instead it uses the lists.h
doubly-linked list implementation, with a list head in the task.c file.

A reference to the kernel task is stored there too. The current task is,
for the moment, and will be moved.

There will now be threads. The thread will contain the context and the
thread api will handle configuring the entry point and stack. Also, need
to look at how the task structures are stored.

At the moment, we're allocating a single virtual page and storing them
all there, but that might cause issues with tasks not in kernel space.

There isn't much of a plan here, I'm sort of making it up as I go along.